Add nspooler support
We now have a non-standard pooler service
https://nspooler-service-prod-1.delivery.puppetlabs.net/api/v1
but no longer a tool for interacting with it on the command line.

This patch tacks on an nspooler command to floaty that offers basic
support for listing nspooler resources, reserving, querying and
releasing hosts.

It does not try to make the base floaty commands work smoothly across
both poolers.